feat: Add the overview cta to create a new agent
